Remove /game command and allow battles without class
