+++ discord-bot/deploy-commands.js
@@ -9,7 +9,6 @@ const commandDirs = [
   path.join(__dirname, 'commands/admin.js'),
   path.join(__dirname, 'commands/inventory.js'),
   path.join(__dirname, 'commands/set.js'),
-  path.join(__dirname, 'src/commands/game.js'),
   path.join(__dirname, 'src/commands/adventure.js'),
   path.join(__dirname, 'src/commands/challenge.js'),
   path.join(__dirname, 'src/commands/practice.js')
